centralize edit market price
