+++ src/pages/trade/disposable.tsx
@@ -12,13 +12,11 @@ import { getDefaultOrder } from 'components/strategies/create/utils';
 import { StrategyChartHistory } from 'components/strategies/common/StrategyChartHistory';
 import { StrategyChartSection } from 'components/strategies/common/StrategyChartSection';
 import { useTradeCtx } from 'components/trade/TradeContext';
-import { TradeLayout } from 'components/trade/TradeLayout';
 import { useMarketPrice } from 'hooks/useMarketPrice';
 import { StrategyDirection } from 'libs/routing';
 import { TradeDisposableSearch } from 'libs/routing/routes/trade';
 import { useCallback } from 'react';
-import { CarbonLogoLoading } from 'components/common/CarbonLogoLoading';
-import { OverlappingInitMarketPrice } from 'components/strategies/overlapping/OverlappingMarketPrice';
+import { CreateLayout } from 'components/strategies/create/CreateLayout';
 
 const url = '/trade/disposable';
 export const TradeDisposable = () => {
@@ -70,58 +68,9 @@ export const TradeDisposable = () => {
     sell: order.settings !== 'range',
   };
 
-  const setMarketPrice = useCallback(
-    (marketPrice: string) => {
-      navigate({
-        search: (previous) => ({ ...previous, marketPrice }),
-        replace: true,
-        resetScroll: false,
-      });
-    },
-    [navigate]
-  );
-
-  if (!marketPrice && marketQuery.isPending) {
-    return (
-      <TradeLayout>
-        <CarbonLogoLoading className="h-[80px] place-self-center" />
-      </TradeLayout>
-    );
-  }
-
-  if (!marketPrice) {
-    return (
-      <>
-        <TradeLayout>
-          <article
-            key="marketPrice"
-            className="bg-background-900 rounded-ee rounded-es"
-          >
-            <OverlappingInitMarketPrice
-              base={base}
-              quote={quote}
-              setMarketPrice={(price) => setMarketPrice(price)}
-            />
-          </article>
-        </TradeLayout>
-        <StrategyChartSection>
-          <StrategyChartHistory
-            type="recurring"
-            base={base}
-            quote={quote}
-            order0={order0}
-            order1={order1}
-            isLimit={isLimit}
-            onPriceUpdates={onPriceUpdates}
-          />
-        </StrategyChartSection>
-      </>
-    );
-  }
-
   return (
     <>
-      <TradeLayout>
+      <CreateLayout url={url}>
         <CreateForm base={base} quote={quote} order0={order0} order1={order1}>
           <CreateOrder
             type="disposable"
@@ -153,7 +102,7 @@ export const TradeDisposable = () => {
             }
           />
         </CreateForm>
-      </TradeLayout>
+      </CreateLayout>
       <StrategyChartSection>
         <StrategyChartHistory
           type="disposable"
