+++ src/pages/trade/recurring.tsx
@@ -1,4 +1,4 @@
-import { useNavigate, useSearch } from '@tanstack/react-router';
+import { useSearch } from '@tanstack/react-router';
 import { useSetRecurringOrder } from 'components/strategies/common/useSetOrder';
 import { outSideMarketWarning } from 'components/strategies/common/utils';
 import { CreateOrder } from 'components/strategies/create/CreateOrder';
@@ -15,13 +15,10 @@ import { TradeLayout } from 'components/trade/TradeLayout';
 import { StrategyChartHistory } from 'components/strategies/common/StrategyChartHistory';
 import { useCallback } from 'react';
 import { OnPriceUpdates } from 'components/strategies/common/d3Chart';
-import { OverlappingInitMarketPrice } from 'components/strategies/overlapping/OverlappingMarketPrice';
-import { CarbonLogoLoading } from 'components/common/CarbonLogoLoading';
 
 const url = '/trade/recurring';
 export const TradeRecurring = () => {
   const search = useSearch({ from: url });
-  const navigate = useNavigate({ from: url });
   const { base, quote } = useTradeCtx();
   const { setSellOrder, setBuyOrder } = useSetRecurringOrder(url);
   const marketQuery = useMarketPrice({ base, quote });
@@ -74,55 +71,6 @@ export const TradeRecurring = () => {
     buy: buyOrder.settings === 'limit',
     sell: sellOrder.settings === 'limit',
   };
-
-  const setMarketPrice = useCallback(
-    (marketPrice: string) => {
-      navigate({
-        search: (previous) => ({ ...previous, marketPrice }),
-        replace: true,
-        resetScroll: false,
-      });
-    },
-    [navigate]
-  );
-
-  if (!marketPrice && marketQuery.isPending) {
-    return (
-      <TradeLayout>
-        <CarbonLogoLoading className="h-[80px] place-self-center" />
-      </TradeLayout>
-    );
-  }
-
-  if (!marketPrice) {
-    return (
-      <>
-        <TradeLayout>
-          <article
-            key="marketPrice"
-            className="bg-background-900 rounded-ee rounded-es"
-          >
-            <OverlappingInitMarketPrice
-              base={base}
-              quote={quote}
-              setMarketPrice={(price) => setMarketPrice(price)}
-            />
-          </article>
-        </TradeLayout>
-        <StrategyChartSection>
-          <StrategyChartHistory
-            type="recurring"
-            base={base}
-            quote={quote}
-            order0={buyOrder}
-            order1={sellOrder}
-            isLimit={isLimit}
-            onPriceUpdates={onPriceUpdates}
-          />
-        </StrategyChartSection>
-      </>
-    );
-  }
 
   return (
     <>
